# src/server/interface/generated/swagger.yaml
components:
    examples: {}
    headers: {}
    parameters: {}
    requestBodies: {}
    responses: {}
    schemas:
        Identifiable:
            properties:
                id:
                    type: string
            required:
                - id
            type: object
            additionalProperties: false
        JsonCategory:
            properties:
                description:
                    type: string
                    nullable: true
                logoUrlStr:
                    type: string
                    nullable: true
                name:
                    type: string
            required:
                - name
            type: object
            additionalProperties: false
        IdentifiableCategory:
            properties:
                description:
                    type: string
                    nullable: true
                logoUrlStr:
                    type: string
                    nullable: true
                name:
                    type: string
                id:
                    type: string
            required:
                - name
                - id
            type: object
            additionalProperties: false
        IdentifiablePerson:
            properties:
                id:
                    type: string
                avatarUrlStr:
                    type: string
                    nullable: true
                prename:
                    type: string
                surname:
                    type: string
            required:
                - id
                - prename
                - surname
            type: object
            additionalProperties: false
        Retailer:
            properties:
                logoUrlStr:
                    type: string
                    nullable: true
                name:
                    type: string
            required:
                - name
            type: object
            additionalProperties: false
        IdentifiableRetailer:
            properties:
                id:
                    type: string
                logoUrlStr:
                    type: string
                    nullable: true
                name:
                    type: string
            required:
                - id
                - name
            type: object
            additionalProperties: false
    securitySchemes: {}
info:
    title: TSOA
    version: 1.0.0
    description: 'Build swagger-compliant REST APIs using TypeScript and Node'
    license:
        name: MIT
openapi: 3.0.0
paths:
    /categories:
        post:
            operationId: CreateCategory
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Identifiable'
                    description: Created
                '400':
                    content:
                        application/json: {}
                    description: 'Bad Request'
            security: []
            parameters: []
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/JsonCategory'
        get:
            operationId: ReadCategories
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Identifiable'
                                type: array
                    description: OK
            security: []
            parameters: []
    '/categories/{id}':
        get:
            operationId: ReadCategory
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IdentifiableCategory'
                    description: OK
                '404':
                    content:
                        application/json: {}
                    description: 'Not Found'
            security: []
            parameters:
                -
                    in: path
                    name: id
                    required: true
                    schema:
                        type: string
    /persons:
        get:
            operationId: GetPersonsSync
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/IdentifiablePerson'
                                type: array
                    description: Ok
            security: []
            parameters: []
    '/persons/{id}':
        get:
            operationId: GetPersonSync
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IdentifiablePerson'
                    description: Ok
            security: []
            parameters:
                -
                    in: path
                    name: id
                    required: true
                    schema:
                        type: string
    /retailers:
        post:
            operationId: CreateRetailer
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Identifiable'
                    description: Created
                '400':
                    content:
                        application/json: {}
                    description: 'Bad Request'
            security: []
            parameters: []
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Retailer'
        get:
            operationId: ReadRetailers
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Identifiable'
                                type: array
                    description: OK
            security: []
            parameters: []
    '/retailers/{id}':
        get:
            operationId: ReadRetailer
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IdentifiableRetailer'
                    description: OK
                '404':
                    content:
                        application/json: {}
                    description: 'Not Found'
            security: []
            parameters:
                -
                    in: path
                    name: id
                    required: true
                    schema:
                        type: string
servers:
    -
        url: 'https://localhost:8080/v0'